Split the listeners in two
