+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
  - Add the `route` tag automatically to any event generated by a request (#167, thanks @ruudk)
 ### Changed
- - Make the `ExceptionListener` more extendable by making all members at least `protected` (#176, thanks @bouland)
+ - Make the `RequestListener` more extendable by making all members at least `protected` (#176, thanks @bouland)
 
 ## 2.1.0 - 2018-11-05
 ### Added
@@ -66,8 +66,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Add official support to PHP 7.2 (#71)
 ### Changed
  - Changed source folder from `src/Sentry/SentryBundle` to just `src/` (thanks to PSR-4 and Composer this doesn't affect you)
- - Re-sort the constructor's arguments of `ExceptionListener`
- - The `SentrySymfonyClient` is no longer an optional argument of `ExceptionListener`; it's now required
+ - Re-sort the constructor's arguments of `RequestListener`
+ - The `SentrySymfonyClient` is no longer an optional argument of `RequestListener`; it's now required
 ### Fixed
  - Remove usage of `create_function()` to avoid deprecations (#71)
  - Fix a possible bug that could make Sentry crash if an error is triggered before loading a console command
