+++ test/Fixtures/InvalidExceptionListener.php
@@ -3,7 +3,7 @@
 namespace Sentry\SentryBundle\Test\Fixtures;
 
 /**
- * An invalid <code>ExceptionListener</code> for testing.
+ * An invalid <code>RequestListener</code> for testing.
  *
  * This exception listener does not implement, or extend a class that
  * implements, the appropriate interface that the dependency injection
